made time utility
